requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
httpx==0.26.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1